wire(hello): emit connectionId as int32 to match MongoDB 7.0 (closes #744)
PR #743 emitted connectionId as int64. Verified against MongoDB Server r7.0.0
source (src/mongo/db/repl/replication_info.cpp + bson/bsonobjbuilder.h):
hello uses BSONObjBuilder::appendNumber(StringData, long long), which
dispatches to int32 when the value fits in [INT_MIN, INT_MAX]. Connection
IDs are monotonically incrementing per-mongod (atomic.Int64 from zero in
salvobase too) and will fit in int32 in any realistic scenario — ~25 days
at 10k conn/s to overflow.

Change is one line in hello.go plus matching fixture updates in the bench
tests so TestHandleHello_WireEquivalent continues to pin the int32 type
end-to-end. Bench ratio between Legacy and Template is unaffected (both
shrink by the same 4 bytes).

+++ internal/commands/marshal_bench_test.go
@@ -23,7 +23,7 @@ func helloResponseDoc() bson.D {
 		{Key: "maxWriteBatchSize", Value: int32(100000)},
 		{Key: "localTime", Value: bson.DateTime(now.UnixMilli())},
 		{Key: "logicalSessionTimeoutMinutes", Value: int32(30)},
-		{Key: "connectionId", Value: int64(42)},
+		{Key: "connectionId", Value: int32(42)},
 		{Key: "minWireVersion", Value: int32(0)},
 		{Key: "maxWireVersion", Value: int32(21)},
 		{Key: "readOnly", Value: false},
